+++ Project/Assets/Scripts/Generics/Health.cs
@@ -4,7 +4,7 @@ using System.Collections;
 
 abstract public class Health : MonoBehaviour
 {
-    abstract public void TakeDamage(int amount);
+    abstract public void TakeDamage(float amount);
     abstract public void Death();
 }
